handle up to 2^64 - 1 by using long double

+++ include/BinaryOpNode.h
@@ -19,7 +19,7 @@ public:
     AddNode(unique_ptr<Node> left, unique_ptr<Node> right)
         : child1(move(left)), child2(move(right)) {}
 
-    double evaluate(const map<string, double>& variables) const override {
+    long double evaluate(const map<string, long double>& variables) const override {
         return child1->evaluate(variables) + child2->evaluate(variables);
     }
     Node* clone() const override {
@@ -36,7 +36,7 @@ public:
     SubtractNode(unique_ptr<Node> left, unique_ptr<Node> right)
         : child1(move(left)), child2(move(right)) {}
 
-    double evaluate(const map<string, double>& variables) const override {
+    long double evaluate(const map<string, long double>& variables) const override {
         return child1->evaluate(variables) - child2->evaluate(variables);
     }
     Node* clone() const override {
@@ -53,7 +53,7 @@ public:
     MultiplyNode(unique_ptr<Node> left, unique_ptr<Node> right)
         : child1(move(left)), child2(move(right)) {}
 
-    double evaluate(const map<string, double>& variables) const override {
+    long double evaluate(const map<string, long double>& variables) const override {
         return child1->evaluate(variables) * child2->evaluate(variables);
     }
     Node* clone() const override {
@@ -70,8 +70,8 @@ public:
     DivideNode(unique_ptr<Node> left, unique_ptr<Node> right)
         : numerator(move(left)), denominator(move(right)) {}
 
-    double evaluate(const map<string, double>& variables) const override {
-        double denom = denominator->evaluate(variables);
+    long double evaluate(const map<string, long double>& variables) const override {
+        long double denom = denominator->evaluate(variables);
         if (denom == 0) {
             throw runtime_error("Division by zero error");
         }
@@ -91,7 +91,7 @@ public:
     PowerNode(unique_ptr<Node> base, unique_ptr<Node> exponent)
         : base(move(base)), exponent(move(exponent)) {}
 
-    double evaluate(const map<string, double>& variables) const override {
+    long double evaluate(const map<string, long double>& variables) const override {
         if (base->evaluate(variables) < 0 && exponent->evaluate(variables) != floor(exponent->evaluate(variables))) {
             throw runtime_error("Invalid operation: negative base with non-integer exponent");
         }
